requirements.txt
numpy
Pillow
scipy
six
ImageHash
img2pdf
imutils
pytube
validators
requests